requirements.txt
# ActiBot/requirements.txt
python-dotenv>=0.19.0
# Para el futuro con FastAPI:
# fastapi>=0.70.0
# uvicorn[standard]>=0.15.0
# httpx>=0.20.0
# msal>=1.18.0
# pydantic>=1.9.0